Administrador de Tareas/task_manager/requirements.txt
click==8.1.6
Flask==2.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.1
Werkzeug==2.3.7